requirements.txt
fastapi==0.128.0
uvicorn[standard]==0.40.0
httpx==0.28.1
pydantic==2.12.4
sentence-transformers==5.2.0
torch==2.9.1
sqlite-vec==0.1.6
pyyaml>=6.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
